+++ CHANGELOG.md
@@ -56,7 +56,7 @@ Main feature of that release is full TypeScript support insluding mongoose model
 ## [5.0.0-beta.44]  NEXT
 
 - **[NEW]** Logout method
-
+- **[Update]** ser model typing
 
 ---
 ## [5.0.0-beta.43]